Move RSA tests to separate module
